pyproject.toml
[project]
name = "advanced-seo-mcp"
version = "2.0.0"
description = "Professional SEO Analysis MCP Server for AI Agents"
readme = "README.md"
keywords = ["seo", "mcp", "llm", "ai-agents", "python"]
authors = [
    { name = "Pavel", email = "pavel@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=2.10.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "fake-useragent>=1.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "pre-commit>=3.0",
    "build>=1.2",
]

[project.urls]
Homepage = "https://github.com/kovzun2/advanced-seo-mcp"
Repository = "https://github.com/kovzun2/advanced-seo-mcp"
Issues = "https://github.com/kovzun2/advanced-seo-mcp/issues"

[project.scripts]
advanced-seo = "advanced_seo_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/advanced_seo_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/CONTRIBUTING.md",
    "/pyproject.toml",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra"
